ChangeCulture for numeric + tests

+++ ObjectPrinting/PropertyPrintingConfigExtensions.cs
@@ -26,21 +26,21 @@ namespace ObjectPrinting
         public static PrintingConfig<TOwner> Using<TOwner>(this PropertyPrintingConfig<TOwner, long> propConfig, CultureInfo culture)
         {
             var parentConfig = ((IPropertyPrintingConfig<TOwner, long>)propConfig).ParentConfig;
-            parentConfig.ChangeCultureInfoForType(typeof(int), culture);
+            parentConfig.ChangeCultureInfoForType(typeof(long), culture);
             return parentConfig;
         }
 
         public static PrintingConfig<TOwner> Using<TOwner>(this PropertyPrintingConfig<TOwner, double> propConfig, CultureInfo culture)
         {
             var parentConfig = ((IPropertyPrintingConfig<TOwner, double>)propConfig).ParentConfig;
-            parentConfig.ChangeCultureInfoForType(typeof(int), culture);
+            parentConfig.ChangeCultureInfoForType(typeof(double), culture);
             return parentConfig;
         }
 
         public static PrintingConfig<TOwner> Using<TOwner>(this PropertyPrintingConfig<TOwner, decimal> propConfig, CultureInfo culture)
         {
             var parentConfig = ((IPropertyPrintingConfig<TOwner, decimal>)propConfig).ParentConfig;
-            parentConfig.ChangeCultureInfoForType(typeof(int), culture);
+            parentConfig.ChangeCultureInfoForType(typeof(decimal), culture);
             return parentConfig;
         }
     }
